# swagger.yaml
openapi: 3.0.3
info:
  title: SenAI CRM API
  description: Agentic CRM Intelligence Platform
  version: 1.0.0
  contact:
    name: SenAI CRM Team
servers:
  - url: http://localhost:8000/api
    description: Local development server
paths:
...
  /threads/{contact_email}:
    get:
      summary: Get thread by contact email
      operationId: getThread
      parameters:
        - name: contact_email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Thread retrieved with emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ThreadOut'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /threads/{contact_email}/summary:
    get:
      summary: Get AI-generated thread summary
      operationId: getThreadSummary
      tags: [Threads]
      parameters:
        - name: contact_email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thread summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                  email_count:
                    type: integer
                  generated_by:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error_code
        - message
        - details
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
...
    EmailOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message_id:
          type: string
        sender:
          type: string
        recipient:
          type: string
        subject:
          type: string
        body:
          type: string
        body_truncated:
          type: boolean
        timestamp:
          type: string
          format: date-time
        category:
          type: string
        sentiment_score:
          type: number
        urgency:
          type: string
        confidence:
          type: number
        requires_human:
          type: boolean
        status:
          type: string
        raw_entities:
          type: object
        classification_result:
          type: object
        created_at:
          type: string
          format: date-time
    ThreadOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_email:
          type: string
        subject:
          type: string
        category:
          type: string
        status:
          type: string
        last_updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        emails:
          type: array
          items:
            $ref: '#/components/schemas/EmailOut'
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error_code: NOT_FOUND
            message: Resource not found
            details: {}
...